@@ -64,18 +64,17 @@ final class ServiceLocatorTagPass extends AbstractRecursivePass
             if ($v instanceof ServiceClosureArgument) {
                 continue;
             }
-            if (!$v instanceof Reference) {
-                throw new InvalidArgumentException(sprintf('Invalid definition for service "%s": an array of references is expected as first argument when the "container.service_locator" tag is set, "%s" found for key "%s".', $this->currentId, get_debug_type($v), $k));
-            }
 
             if ($i === $k) {
-                unset($services[$k]);
-
-                $k = (string) $v;
+                if ($v instanceof Reference) {
+                    unset($services[$k]);
+                    $k = (string) $v;
+                }
                 ++$i;
             } elseif (\is_int($k)) {
                 $i = null;
             }
+
             $services[$k] = new ServiceClosureArgument($v);
         }
         ksort($services);
@@ -97,20 +96,14 @@ final class ServiceLocatorTagPass extends AbstractRecursivePass
         return new Reference($id);
     }
 
-    /**
-     * @param Reference[] $refMap
-     */
-    public static function register(ContainerBuilder $container, array $refMap, string $callerId = null): Reference
+    public static function register(ContainerBuilder $container, array $map, string $callerId = null): Reference
     {
-        foreach ($refMap as $id => $ref) {
-            if (!$ref instanceof Reference) {
-                throw new InvalidArgumentException(sprintf('Invalid service locator definition: only services can be referenced, "%s" found for key "%s". Inject parameter values using constructors instead.', get_debug_type($ref), $id));
-            }
-            $refMap[$id] = new ServiceClosureArgument($ref);
+        foreach ($map as $k => $v) {
+            $map[$k] = new ServiceClosureArgument($v);
         }
 
         $locator = (new Definition(ServiceLocator::class))
-            ->addArgument($refMap)
+            ->addArgument($map)
             ->addTag('container.service_locator');
 
         if (null !== $callerId && $container->hasDefinition($callerId)) {
