@@ -24,7 +24,7 @@ use Symfony\Contracts\Service\ServiceSubscriberInterface;
  * @author Robin Chalas <robin.chalas@gmail.com>
  * @author Nicolas Grekas <p@tchwork.com>
  */
-class ServiceLocator implements ServiceProviderInterface
+class ServiceLocator implements ServiceProviderInterface, \Countable
 {
     use ServiceLocatorTrait {
         get as private doGet;
@@ -74,6 +74,11 @@ class ServiceLocator implements ServiceProviderInterface
         $locator->container = $container;
 
         return $locator;
+    }
+
+    public function count(): int
+    {
+        return \count($this->getProvidedServices());
     }
 
     private function createNotFoundException(string $id): NotFoundExceptionInterface
